sandcastle-ng: make run script show logs
The container is executing "/sbin/init" which has no output, and
prevents us from seeing any logs via the container runtime log
interface.

This change to the run script makes the "podman run" command start the
container in the background, and then proceeds to "exec" journalctl in
the newly running container. Thus allowing us to see the relevant logs
from the system services.